fix(cloud): shared-tier agents navigate views ('go to settings') instead of refusing
Tier-0 shared agents run container-free through runSharedAgentTurn, a bare
generateText/streamText with NO action pipeline and NO plugins. plugin-app-control
(the VIEWS navigation action) never loads there, so 'go to settings', 'go home',
'help', 'change my voice', etc. hit a plain LLM that hallucinates a refusal
('I don't have a settings menu, I'm an AI assistant'). Every new demo user
defaults to a shared agent, so in-app navigation was fully broken for them.

Root-cause fix, no full plugin boot:
- Extract the deterministic multilingual view-command matcher into
  @elizaos/shared/views/view-command-matcher (plugin-app-control now re-exports
  it verbatim), so the shared runtime resolves navigation intents from the SAME
  source as the dedicated VIEWS action.
- shared-nav-intent.ts: resolve a nav command -> builtin view id (+ Settings
  'voice' deep-link for 'change my voice'); runSharedAgentTurn(Stream) short-
  circuits with a canned confirmation + navIntent, NO model call, NO billing.
- eliza-sandbox bridgeSharedMessageSend/Stream attach a VIEWS handoff
  (actionResults:[{actionName:VIEWS,success:true,values:{mode:show,viewId,subview?}}])
  to the reply/done frame — the exact shape findViewActionHandoff already reads.
- Client: shared/limited cloud agents serve no /api/views/current, so route the
  handoff through dispatchViewActionHandoffDirect (dispatch straight from the
  summary, no verify round-trip) instead of throwing on the missing route.

Precision-first: only unambiguous navigation commands short-circuit; all other
turns fall through to the LLM byte-identical to before. 21 new nav-intent tests
+ 196 matcher parity tests green. 5 canonical commands verified end-to-end.

Lane F3 (sol-f5-actions). Repro + fix receipts in memory/projects/eliza-fleet.
